+++ CHANGELOG.md
@@ -1,8 +1,26 @@
+# BlazingSQL 0.20.0 (Date TBS)
+
+
+## New Features
+
+
+
+## Improvements
+
+
+
+## Bug Fixes
+
+
+
 # BlazingSQL 0.19.0 (Date TBS)
 
 ## New Features
+- #1367 OverlapAccumulator Kernel
 - #1364 Implement the concurrent API (bc.sql with token, bc.status, bc.fetch)
+- #1426 Window Functions without partitioning 
 - #1349 Add e2e test for Hive Partitioned Data
+- #1396 Create tables from other RDBMS
 - #1427 Support for CONCAT alias operator
 - #1424 Add get physical plan with explain
 
@@ -14,7 +32,9 @@
 - #1376 Fixing build due to some strings refactor in cudf, undoing the replace workaround
 - #1331 Added flag to enable null e2e testing
 - #1418 Adding support for docker image
+- #1434 Added documentation for C++ and Python in Sphinx
 - #1419 Added concat cache machine timeout 
+- #1349 Add e2e test for Hive Partitioned Data
 - #1447 Improve getting estimated output num rows
 
 ## Bug Fixes
@@ -43,7 +63,7 @@
 - #1351 Fixed 'count distinct' related issues
 - #1425 Fix for new joins API
 - #1400 Fix for Column aliases when exists a Join op
-
+- #1456 Raising exceptions on Python side for RAL
 
 ## Deprecated Features
 - #1394 Disabled support for outer joins with inequalities